+++ src/pg_upsert/__init__.py
@@ -1,17 +1,14 @@
-from .__version__ import (
-    __author__,
-    __author_email__,
-    __code_url__,
-    __description__,
-    __docs_url__,
-    __license__,
-    __title__,
-    __version__,
-)
+from importlib.metadata import PackageNotFoundError, version
+
 from .cli import app
 from .models import QACheckType, QAError, TableResult, UpsertResult
 from .postgres import PostgresDB
 from .upsert import PgUpsert, UserCancelledError
+
+try:
+    __version__ = version("pg_upsert")
+except PackageNotFoundError:
+    __version__ = "unknown"
 
 __all__ = [
     "PgUpsert",