+++ src/pg_upsert/__main__.py
@@ -1,5 +1,4 @@
-from .__version__ import __title__
 from .cli import app
 
 if __name__ == "__main__":
-    app(prog_name=__title__)
+    app(prog_name="pg_upsert")